refactor: add movement and command handling functions for player actions
