+++ server/src/functional/handle_player_ext2.c
@@ -117,103 +117,3 @@ void move_player(client_t *client, server_config_t *cnf, direction_t direction)
     if (direction >= 0 && direction < 4)
         move_functions[direction](client, cnf);
 }
-
-void execute_command(client_t *client, server_config_t *conf, const char *cmd)
-{
-    char inventory[256];
-    int eject_direction;
-    char *message;
-    int direction_tile;
-    char broadcast_msg[512];
-    char response[32];
-
-    if (strncmp(cmd, "Forward", 7) == 0) {
-        move_player(client, conf, client->direction);
-        send(client->fd, "ok\n", 3, 0);
-        ppo_graphics(client, conf);
-        pin_graphics(client, conf);
-        return;
-    }
-    if (strncmp(cmd, "Right", 5) == 0) {
-        client->direction = (client->direction + 1) % 4;
-        send(client->fd, "ok\n", 3, 0);
-        ppo_graphics(client, conf);
-        pin_graphics(client, conf);
-        return;
-    }
-    if (strncmp(cmd, "Left", 4) == 0) {
-        client->direction = (client->direction + 3) % 4;
-        send(client->fd, "ok\n", 3, 0);
-        ppo_graphics(client, conf);
-        pin_graphics(client, conf);
-        return;
-    }
-    if (strncmp(cmd, "Inventory", 9) == 0) {
-        get_tile_contents(client, inventory);
-        snprintf(inventory, sizeof(inventory), "[linemate %d, deraumere %d, "
-        "sibur %d, mendiane %d, phiras %d, thystame %d, food %d]\n",
-            client->inventory.linemate,
-            client->inventory.deraumere,
-            client->inventory.sibur,
-            client->inventory.mendiane,
-            client->inventory.phiras,
-            client->inventory.thystame,
-            client->inventory.food
-        );
-        send(client->fd, inventory, strlen(inventory), 0);
-        return;
-    }
-    if (strncmp(cmd, "Take", 4) == 0) {
-        take_object(client, conf, (char *)cmd);
-        pin_graphics(client, conf);
-        return;
-    }
-    if (strncmp(cmd, "Look", 4) == 0) {
-        look_around(client, conf);
-        return;
-    }
-    if (strncmp(cmd, "Fork", 4) == 0) {
-        create_egg_fork(conf, client);
-        send(client->fd, "ok\n", 3, 0);
-        return;
-    }
-    if (strncmp(cmd, "Broadcast", 9) == 0) {
-        message = (char *)cmd + 10;
-        for (client_t *c = conf->clients; c != NULL; c = c->next) {
-            if (c != client) {
-                direction_tile = calculate_direction_tile(client, c);
-                snprintf(broadcast_msg, sizeof(broadcast_msg),
-                    "message %d, %s\n", direction_tile, message);
-                send(c->fd, broadcast_msg, strlen(broadcast_msg), 0);
-            }
-        }
-        send(client->fd, "ok\n", 3, 0);
-        return;
-    }
-    if (strncmp(cmd, "Connect_nbr", 11) == 0) {
-        snprintf(response, sizeof(response), "%d\n", conf->team_count *
-            conf->clients_nb - find_nb_teams(client) + 1);
-        send(client->fd, response, strlen(response), 0);
-        return;
-    }
-    if (strncmp(cmd, "Set", 3) == 0) {
-        set_object(client, conf, (char *)cmd, client->fd);
-        return;
-    }
-    if (strncmp(cmd, "Eject", 5) == 0) {
-        for (client_t *c = conf->clients; c != NULL; c = c->next) {
-            if (c != client && c->x == client->x && c->y == client->y) {
-                eject_direction = (client->direction + 2) % 4;
-                move_player(c, conf, eject_direction);
-                notify_graphics_player_update(c, conf);
-            }
-        }
-        send(client->fd, "ok\n", 3, 0);
-        return;
-    }
-    if (strncmp(cmd, "Incantation", 11) == 0) {
-        handle_incantation(client, conf, client->fd);
-        return;
-    }
-    send(client->fd, "ko\n", 3, 0);
-}
